:sparkles: Add menu. Add settings page. Add Google drive folder select for each category.
