+++ package.json
@@ -17,6 +17,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-firebaseui": "^4.1.0",
+    "react-load-script": "^0.0.6",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
     "react-use": "^15.3.4",
