+++ public/index.html
@@ -17,6 +17,7 @@
     <link rel="manifest" href="%PUBLIC_URL%/manifest.webmanifest" />
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
     <link rel="preload" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
+    <link rel="preload" href="https://apis.google.com/js/api.js" />
     <link rel="preconnect" href="https://www.gstatic.com">
     <!--
       Notice the use of %PUBLIC_URL% in the tags above.
